+++ torchmetrics/__init__.py
@@ -73,6 +73,7 @@ from torchmetrics.regression import (  # noqa: E402
     SpearmanCorrCoef,
     SymmetricMeanAbsolutePercentageError,
     TweedieDevianceScore,
+    WeightedMeanAbsolutePercentageError,
 )
 from torchmetrics.retrieval import (  # noqa: E402
     RetrievalFallOut,
@@ -119,6 +120,7 @@ __all__ = [
     "CalibrationError",
     "CatMetric",
     "ClasswiseWrapper",
+    "CharErrorRate",
     "CHRFScore",
     "CohenKappa",
     "ConfusionMatrix",
@@ -136,6 +138,7 @@ __all__ = [
     "KLDivergence",
     "LabelRankingAveragePrecision",
     "LabelRankingLoss",
+    "MatchErrorRate",
     "MatthewsCorrCoef",
     "MaxMetric",
     "MeanAbsoluteError",
@@ -184,9 +187,8 @@ __all__ = [
     "SymmetricMeanAbsolutePercentageError",
     "TranslationEditRate",
     "UniversalImageQualityIndex",
+    "WeightedMeanAbsolutePercentageError",
     "WordErrorRate",
-    "CharErrorRate",
-    "MatchErrorRate",
     "WordInfoLost",
     "WordInfoPreserved",
 ]