+++ CHANGELOG.md
@@ -58,6 +58,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added new keyword argument `compute_on_cpu` to all metrics ([#867](https://github.com/PyTorchLightning/metrics/pull/867))
 
 
+- Added `WeightedMeanAbsolutePercentageError` to regression package ([#948](https://github.com/PyTorchLightning/metrics/pull/948))
+
+
 ### Changed
 
 - Made `num_classes` in `jaccard_index` a required argument ([#853](https://github.com/PyTorchLightning/metrics/pull/853), [#914](https://github.com/PyTorchLightning/metrics/pull/914))
@@ -134,6 +137,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ### Fixed
+
+- Fixed device mismatch for `MAP` metric in specific cases ([#950](https://github.com/PyTorchLightning/metrics/pull/950))
 
 - Improved testing speed ([#820](https://github.com/PyTorchLightning/metrics/pull/820))